Add generic Git classes
