+++ server/lib/txgh-server/webhooks.rb
@@ -1,5 +1,6 @@
 module TxghServer
   module Webhooks
+    autoload :Git,       'txgh-server/webhooks/git'
     autoload :Github,    'txgh-server/webhooks/github'
     autoload :Gitlab,    'txgh-server/webhooks/gitlab'
     autoload :Transifex, 'txgh-server/webhooks/transifex'
