+++ server/lib/txgh-server/webhooks/github/blank_attributes.rb
@@ -1,19 +1,7 @@
 module TxghServer
   module Webhooks
     module Github
-      class BlankAttributes
-        class << self
-          def from_webhook_payload(payload)
-            new
-          end
-        end
-
-        def initialize(options = {})
-        end
-
-        def to_h
-          {}
-        end
+      class BlankAttributes < TxghServer::Webhooks::Git::BlankAttributes
       end
     end
   end
