+++ server/lib/txgh-server/webhooks/github/ping_handler.rb
@@ -3,18 +3,7 @@ module TxghServer
     module Github
       # Handles github's ping event, which is a test event fired whenever a new
       # webhook is set up.
-      class PingHandler
-        include ResponseHelpers
-
-        attr_reader :logger
-
-        def initialize(logger)
-          @logger = logger
-        end
-
-        def execute
-          respond_with(200, {})
-        end
+      class PingHandler < TxghServer::Webhooks::Git::PingHandler
       end
     end
   end
