+++ server/lib/txgh-server/webhooks/github/push_attributes.rb
@@ -1,39 +1,10 @@
 module TxghServer
   module Webhooks
     module Github
-      class PushAttributes
-        ATTRIBUTES = [
-          :event, :repo_name, :ref, :before, :after,
-          :added_files, :modified_files, :author
-        ]
-
+      class PushAttributes < TxghServer::Webhooks::Git::PushAttributes
         class << self
-          def from_webhook_payload(payload)
-            new(
-              ATTRIBUTES.each_with_object({}) do |attr, ret|
-                ret[attr] = public_send(attr, payload)
-              end
-            )
-          end
-
-          def event(payload)
-            'push'
-          end
-
           def repo_name(payload)
             payload.fetch('repository').fetch('full_name')
-          end
-
-          def ref(payload)
-            payload.fetch('ref')
-          end
-
-          def before(payload)
-            payload.fetch('before')
-          end
-
-          def after(payload)
-            payload.fetch('after')
           end
 
           def added_files(payload)
@@ -57,27 +28,6 @@ module TxghServer
             payload.fetch('commits').flat_map { |c| c[state] }.uniq
           end
         end
-
-        def files
-          @files ||= added_files + modified_files
-        end
-
-        attr_reader *ATTRIBUTES
-
-        def initialize(options = {})
-          ATTRIBUTES.each do |attr|
-            instance_variable_set(
-              "@#{attr}", options.fetch(attr) { options.fetch(attr.to_s) }
-            )
-          end
-        end
-
-        def to_h
-          ATTRIBUTES.each_with_object({}) do |attr, ret|
-            ret[attr] = public_send(attr)
-          end
-        end
-
       end
     end
   end
