+++ server/lib/txgh-server/webhooks/github/request_handler.rb
@@ -3,22 +3,7 @@ require 'json'
 module TxghServer
   module Webhooks
     module Github
-      class RequestHandler
-        include ResponseHelpers
-
-        class << self
-          def handle_request(request, logger)
-            new(request, logger).handle_request
-          end
-        end
-
-        attr_reader :request, :logger
-
-        def initialize(request, logger)
-          @request = request
-          @logger = logger
-        end
-
+      class RequestHandler < TxghServer::Webhooks::Git::RequestHandler
         def handle_request
           handle_safely do
             case github_event
@@ -47,42 +32,12 @@ module TxghServer
           end
         end
 
-        def handle_safely
-          if authentic_request?
-            yield
-          else
-            respond_with_error(401, 'Unauthorized')
-          end
-        rescue => e
-          respond_with_error(500, "Internal server error: #{e.message}", e)
-        end
-
-        def payload
-          @payload ||= begin
-            if request.params[:payload]
-              JSON.parse(request.params[:payload])
-            else
-              JSON.parse(request.body.read)
-            end
-          rescue JSON::ParserError
-            {}
-          end
-        end
-
         def github_event
           request.env['HTTP_X_GITHUB_EVENT']
         end
 
         def git_repo_name
           payload.fetch('repository', {})['full_name']
-        end
-
-        def config
-          @config ||= Txgh::Config::KeyManager.config_from_repo(git_repo_name)
-        end
-
-        def repo
-          config.git_repo
         end
 
         def project
