+++ server/lib/txgh-server/webhooks/gitlab/blank_attributes.rb
@@ -1,7 +1,7 @@
 module TxghServer
   module Webhooks
     module Gitlab
-      class BlankAttributes < TxghServer::Webhooks::Github::BlankAttributes
+      class BlankAttributes < TxghServer::Webhooks::Git::BlankAttributes
       end
     end
   end
