+++ server/lib/txgh-server/webhooks/gitlab/delete_handler.rb
@@ -1,12 +1,7 @@
 module TxghServer
   module Webhooks
     module Gitlab
-      class DeleteHandler < TxghServer::Webhooks::Github::DeleteHandler
-        private
-
-        def should_handle_request?
-          repo.should_process_ref?(branch) && project.auto_delete_resources?
-        end
+      class DeleteHandler < TxghServer::Webhooks::Git::DeleteHandler
       end
     end
   end
